Update to latest versions of everything.
Rust 2018 and Rocket 0.4 make a lot of changes that I personally have
been waiting for, for a long time. The stdlib Error trait is currently being
worked on, so in the future we might not even need the failure crate.
Due to the changes to Rocket I have removed slog completely and it's
much nicer without it.